international tables created, update Datasets.zip

+++ Scripts/createCountriesTable.py
@@ -1,6 +1,7 @@
 #TABLE NAME: countries
 #csv file: countries.csv
 import csv
+import os
 def create_countries_table(connection):
     # create a cursor
     cur = connection.cursor()
@@ -38,26 +39,40 @@ def create_countries_table(connection):
     
     countries_names_dict = {}
     countries_fips_dict = {}
-    with open('../Datasets/countries/countries.csv', 'r', encoding='iso-8859-1') as infile:
-        with open('../Datasets/countries/countries_updated.csv', 'w+', encoding='utf-8') as outfile:
+
+    if os.path.exists('../Datasets/countries/countries_updated.csv'):
+        with open('../Datasets/countries/countries_updated.csv', 'r', encoding='utf-8') as infile:
             reader = csv.reader(infile)
-            writer = csv.writer(outfile)
             for row in reader:
-                for i in range(len(row)):
-                    if "," in row[i]:
-                        row[i] = row[i].replace(",", "")
-                    elif "#N/A" in row[i]:
-                        row[i] = ""
-                    
                 countries_names_dict[row[5]] = row[2]
                 if row [3] != '':
                     countries_fips_dict[row[3]] = row[2]
                 else:
                     countries_fips_dict[row[0]] = row[2]
-                writer.writerow(row)   
-            outfile.seek(0) # move the file pointer to the beginning of the file         
-            next(outfile) # skip header row
-            #cur.copy_from(outfile, 'countries', sep=',', null='')
+            infile.seek(0) # move the file pointer to the beginning of the file
+            next(infile) # skip header row
+            cur.copy_from(infile, 'countries', sep=',', null='')
+    else:
+        with open('../Datasets/countries/countries.csv', 'r', encoding='iso-8859-1') as infile:
+            with open('../Datasets/countries/countries_updated.csv', 'w+', encoding='utf-8') as outfile:
+                reader = csv.reader(infile)
+                writer = csv.writer(outfile)
+                for row in reader:
+                    for i in range(len(row)):
+                        if "," in row[i]:
+                            row[i] = row[i].replace(",", "")
+                        elif "#N/A" in row[i]:
+                            row[i] = ""
+                        
+                    countries_names_dict[row[5]] = row[2]
+                    if row [3] != '':
+                        countries_fips_dict[row[3]] = row[2]
+                    else:
+                        countries_fips_dict[row[0]] = row[2]
+                    writer.writerow(row)   
+                outfile.seek(0) # move the file pointer to the beginning of the file         
+                next(outfile) # skip header row
+                cur.copy_from(outfile, 'countries', sep=',', null='')
 
     # close the cursor and connection
     cur.close()
